+++ ReportPrinter/ReportPrinterUnitTest/RaphaelLibrary/Common/SqlVariableCacheManager/SqlVariableCacheManagerFactoryTest.cs
@@ -10,7 +10,7 @@ namespace ReportPrinterUnitTest.RaphaelLibrary.Common.SqlVariableCacheManager
         [Test]
         [TestCase(CacheManagerType.Memory, typeof(SqlVariableMemoryCacheManager))]
         [TestCase(CacheManagerType.Redis, typeof(SqlVariableRedisCacheManager))]
-        [TestCase(2, null)]
+        [TestCase((byte)2, null)]
         public void TestCreateSqlResultCacheManager(CacheManagerType managerType, Type expectedType)
         {
             try
