+++ ReportPrinter/ReportPrinterUnitTest/RaphaelLibrary/Print/PrinterFactoryTest.cs
@@ -11,7 +11,7 @@ namespace ReportPrinterUnitTest.RaphaelLibrary.Print
         [Test]
         [TestCase(ReportTypeEnum.PDF, typeof(PdfPrinter))]
         [TestCase(ReportTypeEnum.Label, typeof(LabelPrinter))]
-        [TestCase(2, null)]
+        [TestCase((byte)2, null)]
         public void TestCreatePrinter(ReportTypeEnum reportType, Type expectedType)
         {
             try
